#21 MdcProducerDecorator doesn't work with already existing metadata
